+++ python_modules/dagit/dagit/app.py
@@ -17,8 +17,7 @@ from flask_sockets import Sockets
 from graphql.execution.executors.gevent import GeventExecutor as Executor
 from nbconvert import HTMLExporter
 
-from dagster import check, RepositoryDefinition, seven
-from dagster.cli.dynamic_loader import load_repository_from_target_info
+from dagster import check, seven
 from dagster_graphql.schema import create_schema
 
 from dagster_graphql.implementation.context import DagsterGraphQLContext
@@ -29,40 +28,6 @@ from dagster_graphql.implementation.pipeline_execution_manager import (
 from .subscription_server import DagsterSubscriptionServer
 from .templates.playground import TEMPLATE as PLAYGROUND_TEMPLATE
 from .version import __version__
-
-
-class RepositoryContainer(object):
-    '''
-    This class solely exists to implement reloading semantics. We need to have a single object
-    that the graphql server has access that stays the same object between reload. This container
-    object allows the RepositoryInfo to be written in an immutable fashion.
-    '''
-
-    def __init__(self, repository_target_info=None, repository=None):
-        self.repo_error = None
-        if repository_target_info is not None:
-            self.repository_target_info = repository_target_info
-            try:
-                self.repo = check.inst(
-                    load_repository_from_target_info(repository_target_info), RepositoryDefinition
-                )
-            except:  # pylint: disable=W0702
-                self.repo_error = sys.exc_info()
-        elif repository is not None:
-            self.repository_target_info = None
-            self.repo = repository
-
-    @property
-    def repository(self):
-        return self.repo
-
-    @property
-    def error(self):
-        return self.repo_error
-
-    @property
-    def repository_info(self):
-        return self.repository_target_info
 
 
 class DagsterGraphQLView(GraphQLView):
