+++ python_modules/dagit/dagit_tests/test_multiprocessing.py
@@ -6,7 +6,7 @@ from dagster import (
     OutputDefinition,
     lambda_solid,
 )
-from dagster.cli.dynamic_loader import RepositoryTargetInfo
+from dagster.cli.dynamic_loader import RepositoryTargetInfo, RepositoryContainer
 from dagster.core.events import DagsterEventType
 from dagster.core.execution import create_execution_plan, ExecutionSelector
 from dagster.core.execution_context import make_new_run_id
@@ -21,8 +21,6 @@ from dagster_graphql.implementation.pipeline_run_storage import (
 )
 
 import dagster_pandas as dagster_pd
-
-from dagit.app import RepositoryContainer
 
 
 def get_events_of_type(events, event_type):
